requirements.txt
# Flask and Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Session==0.5.0

# Database and Data Processing
duckdb==0.9.2
pandas==2.0.3
numpy==1.24.3

# MS Office Support
python-docx==0.8.11
openpyxl==3.1.2
xlrd==2.0.1

# Web Scraping and APIs
requests==2.31.0
beautifulsoup4==4.12.2
googlesearch-python==1.2.3
wikipedia-api==0.6.0
yfinance==0.2.28

# Utilities
python-dateutil==2.8.2
pytz==2023.3
psutil==5.9.5